fix(#1318): say the install is damaged instead of printing a resolver path
A half-extracted global install produced a bare ERR_MODULE_NOT_FOUND naming a
path inside node_modules — nothing said comfyui-mcp, nothing said the install
was incomplete, and the recorded version still read as current.

src/index.ts becomes a thin launcher that dynamic-imports the real entry
(src/boot.ts) and translates only that failure. It imports nothing from a
package, because a static import is exactly what fails before the handler is
installed.

Diagnoses the observable STATE, not a cause: the reporter was careful that an
interrupted `npm install -g` and a failed self-update are indistinguishable.

Refs #1318
